fix `$PThread` symbol linkage
