+++ packages/emnapi/src/emscripten/async.ts
@@ -1,7 +1,5 @@
-import { ENVIRONMENT_IS_NODE, ENVIRONMENT_IS_PTHREAD } from 'emnapi:emscripten-runtime'
+import { ENVIRONMENT_IS_NODE, ENVIRONMENT_IS_PTHREAD, PThread } from 'emnapi:emscripten-runtime'
 import { _emnapi_set_immediate, _emnapi_next_tick } from '../util'
-
-declare var PThread: any
 
 /**
  * @__deps $PThread
