+++ packages/test/tsfn2/tsfn2_st.test.js
@@ -3,4 +3,4 @@
 const { load } = require('../util')
 const main = require('./main')
 
-module.exports = main(load('tsfn2', { nodeBinding: require('@emnapi/node-binding') }))
+module.exports = main(load('tsfn2_st', { nodeBinding: require('@emnapi/node-binding') }))
